+++ climatoology/app/plugin.py
@@ -10,7 +10,7 @@ from climatoology.app.settings import EXCHANGE_NAME, CABaseSettings, WorkerSetti
 from climatoology.app.tasks import CAPlatformComputeTask
 from climatoology.base.baseoperator import BaseOperator
 from climatoology.base.logging import get_climatoology_logger
-from climatoology.base.plugin_info import PluginInfo
+from climatoology.base.plugin_info import PluginInfoEnriched
 from climatoology.store.database.database import BackendDatabase
 from climatoology.store.database.models.info import PluginInfoTable
 from climatoology.store.object_store import MinioStorage, Storage
@@ -93,7 +93,7 @@ def extract_plugin_id(plugin_id_with_suffix: str) -> str:
     return plugin_id_with_suffix.split('@')[0]
 
 
-def _version_is_compatible(info: PluginInfo, db: BackendDatabase, celery: Celery) -> bool:
+def _version_is_compatible(info: PluginInfoEnriched, db: BackendDatabase, celery: Celery) -> bool:
     with Session(db.engine) as session:
         info_query = session.query(PluginInfoTable).filter_by(id=info.id)
         existing_info = info_query.first()
@@ -123,7 +123,7 @@ def _version_is_compatible(info: PluginInfo, db: BackendDatabase, celery: Celery
     return True
 
 
-def synch_info(info: PluginInfo, db: BackendDatabase, storage: Storage) -> PluginInfo:
+def synch_info(info: PluginInfoEnriched, db: BackendDatabase, storage: Storage) -> PluginInfoEnriched:
     info.assets = storage.write_assets(plugin_id=info.id, assets=info.assets)
 
     _ = db.write_info(info=info)